requirements.txt
opencv_python==4.2.0.32
torch==1.6.0
torchvision==0.7.0
scikit_image==0.16.2
numpy==1.17.4
pycocotools==2.0.2
matplotlib==3.1.2
ipython==7.19.0
Pillow==8.0.1
skimage==0.0
